+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 ## 4.21.0 (2024-08-29)
 
 - added: POL token on Ethereum.
+- added: (Zcash/Piratechain) Log errors thrown by synchronizer
+- added: (Zcash/Piratechain) Restart engine after receiving critical error
 - changed: Rename MATIC to POL on the Polygon chain.
 - fixed: Display name COMBO (Ethereum) to Furucombo
 - fixed: Calculate Arbitrum fees more correctly.